Code style.
