+++ examples/first-component/src/components/Room/createConnection.ts
@@ -1,0 +1,20 @@
+interface createConnectionProps {
+  serverUrl: string;
+  roomId: string;
+}
+
+const createConnection = ({ serverUrl, roomId }: createConnectionProps) => {
+  // A real implementation would actually connect to the server
+  return {
+    connect() {
+      console.log(
+        '✅ Connecting to "' + roomId + '" room at ' + serverUrl + "..."
+      );
+    },
+    disconnect() {
+      console.log('❌ Disconnected from "' + roomId + '" room at ' + serverUrl);
+    },
+  };
+};
+
+export default createConnection;
